libs/pyopengl/requirements.txt
glfw
pyopengl
pyrr
numpy